+++ src/main/java/com/flab/jbly/infrastructure/exception/auth/SessionDoesNotExistException.java
@@ -1,4 +1,4 @@
-package com.flab.jbly.infrastructure.exception;
+package com.flab.jbly.infrastructure.exception.auth;
 
 public class SessionDoesNotExistException extends IllegalArgumentException {
