+++ src/main/java/com/flab/jbly/infrastructure/exception/user/DoesNotAllowLogoutException.java
@@ -1,4 +1,4 @@
-package com.flab.jbly.infrastructure.exception;
+package com.flab.jbly.infrastructure.exception.user;
 
 public class DoesNotAllowLogoutException extends IllegalArgumentException {
