+++ src/main/java/com/flab/jbly/infrastructure/exception/user/NotAllowedUserException.java
@@ -1,4 +1,4 @@
-package com.flab.jbly.infrastructure.exception;
+package com.flab.jbly.infrastructure.exception.user;
 
 public class NotAllowedUserException extends IllegalArgumentException {
